refactor: extract call graph persistence
